+++ include/metal.h
@@ -6,7 +6,7 @@
 class Metal : public Material {
 public:
 	Metal(const Pixel & attenuation);
-	Vec3 scatter(const Ray & r, const gObject * obj, const double t) const;
+	Vec3 scatter(const Ray & r, const gObject * obj, const float t) const;
 };
 
 #endif
