+++ include/pixel.h
@@ -4,22 +4,22 @@
 
 class Pixel {
 private:
-  double r_, g_, b_;
+  float r_, g_, b_;
 public:
-  Pixel(double r = 0, double g = 0, double b = 0): r_{r}, g_{g}, b_{b} {}
+  Pixel(float r = 0, float g = 0, float b = 0): r_{r}, g_{g}, b_{b} {}
   friend std::ostream & operator << (std::ostream & out, const Pixel & pxl);
-  inline double r() const {return r_;}
-  inline double g() const {return g_;}
-  inline double b() const {return b_;}
+  inline float r() const {return r_;}
+  inline float g() const {return g_;}
+  inline float b() const {return b_;}
 
   void operator+=(const Pixel & p2);
-  void operator/=(double c);
+  void operator/=(float c);
 };
 
 Pixel operator+(const Pixel & p1, const Pixel & p2);
-Pixel operator/(const Pixel & p1, double c);
+Pixel operator/(const Pixel & p1, float c);
 Pixel operator*(const Pixel & p1, const Pixel & p2);
-Pixel operator*(const Pixel & p1, double c);
-Pixel operator*(double c, const Pixel & p1);
+Pixel operator*(const Pixel & p1, float c);
+Pixel operator*(float c, const Pixel & p1);
 
 #endif
